fix: revert Qwen3Next scores dtype cast to preserve original behavior
The original code passed float32 scores to _flash_moe (from softmax in
float32). The added .astype(x.dtype) would downcast to bfloat16 in
production — a behavioral regression invisible in tests (which use
float32 inputs).

+++ olmlx/engine/flash/flash_moe_model.py
@@ -113,7 +113,7 @@ class _FlashMoEQwen3Next(_FlashMoEBase):
         scores = mx.take_along_axis(scores, inds, axis=-1)
         if self.norm_topk_prob:
             scores = scores / scores.sum(axis=-1, keepdims=True)
-        return inds, scores.astype(x.dtype)
+        return inds, scores
 
     def _combine(self, x, y):
         shared_y = self.shared_expert(x)
